intro to data.table - JoAnn Rudd Alvarez
**data.table: Where have you been all my life?**

data.table is an easy to use R package for querying and manipulating data. Dataset updates run much faster than with the base R paradigm, and specifying complicated calculations is more intuitive. This talk will show some of the package's features as well as point out a couple of gotchas.